+++ tools/cell_census_builder/src/cell_census_builder/build_soma/validate_soma.py
