+++ tools/cell_census_builder/tests/test_builder.py
@@ -10,7 +10,8 @@ import pyarrow as pa
 import pytest
 import tiledb
 import tiledbsoma as soma
-from cell_census_builder.build_soma.build import build, build_step1_get_source_datasets
+from cell_census_builder.build_soma import build, validate
+from cell_census_builder.build_soma.build_soma import build_step1_get_source_datasets
 from cell_census_builder.build_soma.datasets import Dataset
 from cell_census_builder.build_soma.globals import (
     CENSUS_DATA_NAME,
@@ -18,7 +19,6 @@ from cell_census_builder.build_soma.globals import (
     FEATURE_DATASET_PRESENCE_MATRIX_NAME,
     MEASUREMENT_RNA_NAME,
 )
-from cell_census_builder.build_soma.validate import validate
 from cell_census_builder.build_state import CensusBuildArgs
 
 
@@ -33,10 +33,10 @@ def test_base_builder_creation(
     """
     Runs the builder, queries the census and performs a set of base assertions.
     """
-    with patch("cell_census_builder.build_soma.build.prepare_file_system"), patch(
-        "cell_census_builder.build_soma.build.build_step1_get_source_datasets", return_value=datasets
+    with patch("cell_census_builder.build_soma.build_soma.prepare_file_system"), patch(
+        "cell_census_builder.build_soma.build_soma.build_step1_get_source_datasets", return_value=datasets
     ), patch("cell_census_builder.build_soma.consolidate._run"), patch(
-        "cell_census_builder.build_soma.validate.validate_consolidation", return_value=True
+        "cell_census_builder.build_soma.validate_soma.validate_consolidation", return_value=True
     ):
         return_value = build(census_build_args)
